Temp

+++ backend/src/Squidex.Data.EntityFramework/Domain/Apps/Entities/Contents/EFContentRepository_SnapshotStore.cs
@@ -12,14 +12,16 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Query;
 using Squidex.Domain.Apps.Core.Apps;
 using Squidex.Domain.Apps.Core.Contents;
+using Squidex.Domain.Apps.Core.Schemas;
 using Squidex.Infrastructure;
 using Squidex.Infrastructure.States;
 
 #pragma warning disable MA0048 // File name must match type name
+#pragma warning disable RECS0096 // Type parameter is never used
 
 namespace Squidex.Domain.Apps.Entities.Contents;
 
-public sealed partial class EFContentRepository<TContext> : ISnapshotStore<WriteContent>, IDeleter
+public sealed partial class EFContentRepository<TContext, TContentContext> : ISnapshotStore<WriteContent>, IDeleter
 {
     async IAsyncEnumerable<SnapshotResult<WriteContent>> ISnapshotStore<WriteContent>.ReadAllAsync(
         [EnumeratorCancellation] CancellationToken ct)
@@ -34,8 +36,8 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
         }
     }
 
-    async Task<SnapshotResult<WriteContent>> ISnapshotStore<WriteContent>.ReadAsync(DomainId key,
-        CancellationToken ct)
+    public async Task<SnapshotResult<WriteContent>> ReadAsync(DomainId key,
+        CancellationToken ct = default)
     {
         using (Telemetry.Activities.StartActivity("EFContentRepository/ReadAsync"))
         {
@@ -56,29 +58,28 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
     {
         using (Telemetry.Activities.StartActivity("EFContentRepository/DeleteAppAsync"))
         {
-            await using var dbContext = await CreateDbContextAsync(ct);
-            await using var dbTransaction = await dbContext.Database.BeginTransactionAsync(ct);
-            try
-            {
-                await dbContext.Set<EFContentCompleteEntity>().Where(x => x.IndexedAppId == app.Id)
-                    .ExecuteDeleteAsync(ct);
+            await DeleteCoreAsync(
+                x => x.IndexedAppId == app.Id,
+                x => x.IndexedAppId == app.Id,
+                x => x.AppId == app.Id,
+                x => x.AppId == app.Id,
+                x => x.AppId == app.Id,
+                ct);
+        }
+    }
 
-                await dbContext.Set<EFContentPublishedEntity>().Where(x => x.IndexedAppId == app.Id)
-                    .ExecuteDeleteAsync(ct);
-
-                await dbContext.Set<EFReferenceCompleteEntity>().Where(x => x.AppId == app.Id)
-                    .ExecuteDeleteAsync(ct);
-
-                await dbContext.Set<EFReferencePublishedEntity>().Where(x => x.AppId == app.Id)
-                    .ExecuteDeleteAsync(ct);
-
-                await dbTransaction.CommitAsync(ct);
-            }
-            catch
-            {
-                await dbTransaction.RollbackAsync(ct);
-                throw;
-            }
+    async Task IDeleter.DeleteSchemaAsync(App app, Schema schema,
+        CancellationToken ct)
+    {
+        using (Telemetry.Activities.StartActivity("EFContentRepository/DeleteSchemaAsync"))
+        {
+            await DeleteCoreAsync(
+                x => x.IndexedAppId == app.Id && x.IndexedSchemaId == schema.Id,
+                x => x.IndexedAppId == app.Id && x.IndexedSchemaId == schema.Id,
+                x => x.AppId == app.Id && x.FromSchema == schema.Id,
+                x => x.AppId == app.Id && x.FromSchema == schema.Id,
+                x => x.AppId == app.Id && x.SchemaId == schema.Id,
+                ct);
         }
     }
 
@@ -87,29 +88,60 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
     {
         using (Telemetry.Activities.StartActivity("EFContentRepository/ClearAsync"))
         {
-            await using var dbContext = await CreateDbContextAsync(ct);
-            await using var dbTransaction = await dbContext.Database.BeginTransactionAsync(ct);
-            try
+            await DeleteCoreAsync(
+                x => true,
+                x => true,
+                x => true,
+                x => true,
+                x => true,
+                ct);
+        }
+    }
+
+    private async Task DeleteCoreAsync(
+        Expression<Func<EFContentCompleteEntity, bool>> filterComplete,
+        Expression<Func<EFContentPublishedEntity, bool>> filterPublished,
+        Expression<Func<EFReferenceCompleteEntity, bool>> filterCompleteReferences,
+        Expression<Func<EFReferencePublishedEntity, bool>> filterPublishedReferences,
+        Func<DynamicContextName, bool> filterContexts,
+        CancellationToken ct)
+    {
+        await using var dbContext = await CreateDbContextAsync(ct);
+        await using var dbTransaction = await dbContext.Database.BeginTransactionAsync(ct);
+        try
+        {
+            await dbContext.Set<EFContentCompleteEntity>().Where(filterComplete)
+                .ExecuteDeleteAsync(ct);
+
+            await dbContext.Set<EFContentPublishedEntity>().Where(filterPublished)
+                .ExecuteDeleteAsync(ct);
+
+            await dbContext.Set<EFReferenceCompleteEntity>().Where(filterCompleteReferences)
+                .ExecuteDeleteAsync(ct);
+
+            await dbContext.Set<EFReferencePublishedEntity>().Where(filterPublishedReferences)
+                .ExecuteDeleteAsync(ct);
+
+            if (dedicatedTables)
             {
-                await dbContext.Set<EFContentCompleteEntity>()
-                    .ExecuteDeleteAsync(ct);
+                await foreach (var contextName in dynamicTables.GetContextNames(ct).Where(filterContexts))
+                {
+                    var contentDbContext = await dynamicTables.CreateDbContextAsync(contextName, ct);
 
-                await dbContext.Set<EFContentPublishedEntity>()
-                    .ExecuteDeleteAsync(ct);
+                    await contentDbContext.Set<EFContentCompleteEntity>().Where(filterComplete)
+                        .ExecuteDeleteAsync(ct);
 
-                await dbContext.Set<EFReferenceCompleteEntity>()
-                    .ExecuteDeleteAsync(ct);
-
-                await dbContext.Set<EFReferencePublishedEntity>()
-                    .ExecuteDeleteAsync(ct);
-
-                await dbTransaction.CommitAsync(ct);
+                    await contentDbContext.Set<EFContentPublishedEntity>().Where(filterPublished)
+                        .ExecuteDeleteAsync(ct);
+                }
             }
-            catch
-            {
-                await dbTransaction.RollbackAsync(ct);
-                throw;
-            }
+
+            await dbTransaction.CommitAsync(ct);
+        }
+        catch
+        {
+            await dbTransaction.RollbackAsync(ct);
+            throw;
         }
     }
 
@@ -122,17 +154,20 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
             await using var dbTransaction = await dbContext.Database.BeginTransactionAsync(ct);
             try
             {
-                await dbContext.Set<EFContentCompleteEntity>().Where(x => x.DocumentId == key)
-                    .ExecuteDeleteAsync(ct);
+                if (dedicatedTables)
+                {
+                    var (_, existing, _, _) = await ReadAsync(key, ct);
+                    if (existing == null)
+                    {
+                        return;
+                    }
 
-                await dbContext.Set<EFContentPublishedEntity>().Where(x => x.DocumentId == key)
-                    .ExecuteDeleteAsync(ct);
+                    var contentDbContext = await CreateDbContextAsync(existing.AppId.Id, existing.AppId.Id, ct);
+                    await RemoveCoreAsync(contentDbContext, key, ct);
+                }
 
-                await dbContext.Set<EFReferenceCompleteEntity>().Where(x => x.FromKey == key)
-                    .ExecuteDeleteAsync(ct);
-
-                await dbContext.Set<EFReferencePublishedEntity>().Where(x => x.FromKey == key)
-                    .ExecuteDeleteAsync(ct);
+                // Remove this last, because we need to query the collection.
+                await RemoveCoreAsync(dbContext, key, ct);
 
                 await dbTransaction.CommitAsync(ct);
             }
@@ -142,6 +177,21 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
                 throw;
             }
         }
+    }
+
+    private static async Task RemoveCoreAsync(DbContext dbContext, DomainId key, CancellationToken ct)
+    {
+        await dbContext.Set<EFContentCompleteEntity>().Where(x => x.DocumentId == key)
+            .ExecuteDeleteAsync(ct);
+
+        await dbContext.Set<EFContentPublishedEntity>().Where(x => x.DocumentId == key)
+            .ExecuteDeleteAsync(ct);
+
+        await dbContext.Set<EFReferenceCompleteEntity>().Where(x => x.FromKey == key)
+            .ExecuteDeleteAsync(ct);
+
+        await dbContext.Set<EFReferencePublishedEntity>().Where(x => x.FromKey == key)
+            .ExecuteDeleteAsync(ct);
     }
 
     async Task ISnapshotStore<WriteContent>.WriteAsync(SnapshotWriteJob<WriteContent> job,
@@ -160,25 +210,14 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
 
             try
             {
-                var appId = job.Value.AppId.Id;
+                await AddCoreAsync(dbContext, job, ct);
 
-                await dbContext.Set<EFReferenceCompleteEntity>().Where(x => x.FromKey == job.Key)
-                    .ExecuteDeleteAsync(ct);
-
-                await dbContext.Set<EFReferencePublishedEntity>().Where(x => x.FromKey == job.Key)
-                    .ExecuteDeleteAsync(ct);
-
-                if (job.Value.ShouldWritePublished())
+                if (dedicatedTables)
                 {
-                    await UpsertVersionedPublishedAsync(dbContext, job, ct);
-                }
-                else
-                {
-                    await dbContext.Set<EFContentPublishedEntity>().Where(x => x.DocumentId == job.Key)
-                        .ExecuteDeleteAsync(ct);
+                    var contentContext = await CreateDbContextAsync(job.Value.AppId.Id, job.Value.AppId.Id, ct);
+                    await AddCoreAsync(contentContext, job, ct);
                 }
 
-                await UpsertVersionedCompleteAsync(dbContext, job, ct);
                 await dbTransaction.CommitAsync(ct);
             }
             catch
@@ -187,12 +226,38 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
                 throw;
             }
         }
+
+        async Task AddCoreAsync(DbContext dbContext, SnapshotWriteJob<WriteContent> job, CancellationToken ct)
+        {
+            await dbContext.Set<EFReferenceCompleteEntity>().Where(x => x.FromKey == job.Key)
+                .ExecuteDeleteAsync(ct);
+
+            await dbContext.Set<EFReferencePublishedEntity>().Where(x => x.FromKey == job.Key)
+                .ExecuteDeleteAsync(ct);
+
+            if (job.Value.ShouldWritePublished())
+            {
+                await UpsertVersionedPublishedAsync(dbContext, job, ct);
+            }
+            else
+            {
+                await dbContext.Set<EFContentPublishedEntity>().Where(x => x.DocumentId == job.Key)
+                    .ExecuteDeleteAsync(ct);
+            }
+
+            await UpsertVersionedCompleteAsync(dbContext, job, ct);
+        }
     }
 
     async Task ISnapshotStore<WriteContent>.WriteManyAsync(IEnumerable<SnapshotWriteJob<WriteContent>> jobs,
         CancellationToken ct)
     {
+        // Some data is corrupt and might throw an exception if we do not ignore it.
         var validJobs = jobs.Where(x => IsValid(x.Value)).ToList();
+        if (validJobs.Count == 0)
+        {
+            return;
+        }
 
         using (Telemetry.Activities.StartActivity("EFContentRepository/WriteManyAsync"))
         {
@@ -201,8 +266,6 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
 
             try
             {
-                var keys = validJobs.Select(x => x.Key);
-
                 var writesToCompleteContents = new List<EFContentCompleteEntity>();
                 var writesToCompleteReferences = new List<EFReferenceCompleteEntity>();
                 var writesToPublishedContents = new List<EFContentPublishedEntity>();
@@ -230,8 +293,23 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
                 await dbContext.BulkInsertAsync(writesToCompleteReferences, cancellationToken: ct);
                 await dbContext.BulkInsertAsync(writesToPublishedContents, cancellationToken: ct);
                 await dbContext.BulkInsertAsync(writesToPublishedReferences, cancellationToken: ct);
-
                 await dbContext.SaveChangesAsync(ct);
+
+                if (dedicatedTables)
+                {
+                    foreach (var bySchema in writesToCompleteContents.GroupBy(x => (x.AppId.Id, x.SchemaId.Id)))
+                    {
+                        var (appId, schemaId) = bySchema.Key;
+                        var contentDbContext = await CreateDbContextAsync(appId, schemaId, ct);
+
+                        // Just fetch the published context, so that we can reuse the context.
+                        var publishedContents = writesToPublishedContents.Where(x => x.AppId.Id == appId && x.SchemaId.Id == schemaId);
+
+                        await contentDbContext.BulkInsertAsync(bySchema, cancellationToken: ct);
+                        await contentDbContext.BulkInsertAsync(publishedContents, cancellationToken: ct);
+                    }
+                }
+
                 await dbTransaction.CommitAsync(ct);
             }
             catch
@@ -242,7 +320,7 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
         }
     }
 
-    private async Task UpsertVersionedPublishedAsync(TContext dbContext, SnapshotWriteJob<WriteContent> job,
+    private async Task UpsertVersionedPublishedAsync(DbContext dbContext, SnapshotWriteJob<WriteContent> job,
         CancellationToken ct)
     {
         var (entity, references) = await EFContentPublishedEntity.CreateAsync(job, appProvider, ct);
@@ -251,7 +329,7 @@ public sealed partial class EFContentRepository<TContext> : ISnapshotStore<Write
         await dbContext.UpsertAsync(entity, job.OldVersion, BuildUpdate<EFContentPublishedEntity>, ct);
     }
 
-    private async Task UpsertVersionedCompleteAsync(TContext dbContext, SnapshotWriteJob<WriteContent> job,
+    private async Task UpsertVersionedCompleteAsync(DbContext dbContext, SnapshotWriteJob<WriteContent> job,
         CancellationToken ct)
     {
         var (entity, references) = await EFContentCompleteEntity.CreateAsync(job, appProvider, ct);
